progress on consuming results, but now we overflow ?

+++ packages/converters/models/NodePenDocumentSolutionData.cs
@@ -46,7 +46,7 @@ namespace NodePen.Converters
         /// The number of milliseconds the document took to execute during a given solution.
         /// </summary>
         [JsonProperty("durationMs")]
-        public long DurationMs { get; set; } = 0;
+        public int DurationMs { get; set; } = 0;
 
         public NodePenDocumentRuntimeData()
         {
